Initial work toward adding unit tests
- Migrate from CircleCI 1.0 to 2.0

- Add mocha, chai, sinon, and mock-require as testing framework

- ESLint rule tweaks

- Move assets loaded via Webpack's request.context() into
  src/lib/assets.js for easy mocking in tests outside Webpack

issue #301
