--- circle.yml
@@ -1,26 +1,0 @@
-machine:
-  node:
-    version: 8.9.4
-general:
-  branches:
-    ignore:
-      - gh-pages
-dependencies:
-  override:
-    - npm install -g npm@latest
-    - npm install -g greenkeeper-lockfile@1
-    - greenkeeper-lockfile-update
-    - npm ci
-    - greenkeeper-lockfile-upload
-compile:
-  override:
-    - npm run build 
-deployment:
-  static_development:
-    owner: mozilla
-    branch:
-      - development
-    commands:
-      - git config --global user.email "me@lmorchard.com"
-      - git config --global user.name "Les Orchard"
-      - npm run release:dev
