+++ package.json
@@ -19,7 +19,7 @@
     }
   },
   "scripts": {
-    "start": "npm-run-all --parallel server watch:extension watch:lint watch:validate",
+    "start": "npm-run-all --parallel server watch:extension watch:lint watch:validate watch:test",
     "server": "cross-env NODE_ENV=development SITE_ID=local webpack-dev-server --config webpack.web.js",
     "watch": "npm-run-all --parallel watch:*",
     "watch:web": "cross-env NODE_ENV=development SITE_ID=local webpack --watch --progress --colors --config webpack.web.js",
@@ -27,6 +27,7 @@
     "watch:lint": "npm-run-all --parallel watch:lint:*",
     "watch:lint:js": "onchange -p -v \"src/**/*.js\" -- npm run lint:js",
     "watch:lint:css": "onchange -p -v \"src/**/*.scss\" -- npm run lint:css",
+    "watch:test": "onchange \"src/**/*.js\" -v -i -p -- npm run test",
     "watch:validate": "npm-run-all --parallel watch:validate:*",
     "watch:validate:extension": "onchange -p -v \"build/extension/**/*\" -- npm run validate:extension",
     "watch:validate:themes": "onchange -p -v \"src/preset-themes/*.json\" -- npm run validate:themes",
@@ -45,7 +46,8 @@
     "validate": "npm-run-all validate:*",
     "validate:extension": "addons-linter build/extension --self-hosted",
     "validate:themes": "node ./bin/validate-themes.js",
-    "test": "npm run --parallel lint validate",
+    "test": "npm-run-all --parallel test:*",
+    "test:js": "mocha --require src/lib/test-setup.js --recursive \"src/**/*-test.js\"",
     "deploy": "gh-pages -x -d build/web -r \"https://$GH_TOKEN@github.com/mozilla/FirefoxColor.git\"",
     "release:base": "npm-run-all clean build:web sign && mv addon.xpi build/web && npm run deploy",
     "release:dev": "cross-env ADDON_URL='https://mozilla.github.io/FirefoxColor/addon.xpi' SITE_URL='https://mozilla.github.io/FirefoxColor/' SITE_ID=github npm run release:base"
@@ -70,6 +72,7 @@
     "babel-preset-env": "^1.6.1",
     "babel-preset-node8": "^1.2.0",
     "babel-preset-react": "^6.24.1",
+    "chai": "^4.1.2",
     "copy-webpack-plugin": "^4.4.0",
     "cross-env": "^5.1.3",
     "css-loader": "^0.28.7",
@@ -83,11 +86,16 @@
     "gh-pages": "^1.0.0",
     "html-webpack-plugin": "^3.0.6",
     "image-webpack-loader": "^4.1.0",
+    "mocha": "^5.2.0",
+    "mocha-webpack": "^1.1.0",
+    "mock-require": "^3.0.2",
     "node-sass": "^4.6.1",
     "npm-run-all": "^4.1.2",
     "onchange": "^4.0.0",
+    "require-context": "^1.1.0",
     "rimraf": "^2.6.2",
     "sass-loader": "^7.0.1",
+    "sinon": "^5.0.7",
     "style-loader": "^0.21.0",
     "stylelint": "^9.1.1",
     "stylelint-config-standard": "^18.1.0",
